Change allowBackup to false
Setting allowBackup to true can be a security concern as adb backup allows users who have enabled USB debugging to copy application data off of the device. When working on an app where `allowBackup="false"` is required (such as using visa checkout), Trestle conflicts and the developer must use `tools:replace` to replace the setting in the Trestle manifest. Setting allowBackup to false removes the need to do this.